minor reorgnization
